requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1
python-chess==1.999
python-dotenv==1.0.0
